requirements.txt
greenlet==3.2.3
playwright==1.52.0
pyee==13.0.0
typing_extensions==4.14.0
python-dotenv==1.0.0
langchain-openai
browser-use
